requirements.txt
Flask==1.1.1
gunicorn==19.9.0
python-dotenv==0.10.3
requests==2.22.0
Jinja2==2.10.3
